[favicon] add fetchStrategy service
