+++ src/favicon/fetch-strategies/interfaces/fetch-strategy.interface.ts
@@ -1,0 +1,3 @@
+export interface FetchStrategy {
+  fetchFaviconUrls(subDomainName: string): Promise<Array<string>>;
+}
